// src/models/GRDFDataPoint.ts
export interface GRDFDataPoint {
  dateDebutReleve: string
  dateFinReleve: string
  journeeGaziere: string
  indexDebut: number
  indexFin: number
  volumeBrutConsomme: number
  energieConsomme: number
  pcs: any
  volumeConverti: number
  pta: any
  natureReleve: string
  qualificationReleve: string
  status: any
  coeffConversion: number
  frequenceReleve: any
  temperature: any
}